linkedin-profile-processor/requirements.txt
apache-airflow-providers-amazon>=8.0.0
boto3>=1.26.0
pathlib>=1.0.1
